use integers for id

+++ main-server/database/migrations/20240405203859_create_subscriptions_table.sql
@@ -2,8 +2,8 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS subscriptions
 (
-    follower_id BIGINT NOT NULL REFERENCES users (id),
-    followed_id BIGINT NOT NULL REFERENCES users (id)
+    follower_id INT NOT NULL REFERENCES users (id),
+    followed_id INT NOT NULL REFERENCES users (id)
 );
 -- +goose StatementEnd
